refactor(linters): move generate-types discoverers into the registry (stage 3)
The generate-types `discoverXRules` fns + their config-detection helpers
move from generate-types.ts into linters.ts and attach as each LINTERS
entry's `discoverEnabled` (generateTypes capability now true for all 11).
generate-types.ts drops its parallel `discoverers[]` array and instead maps
`Object.values(LINTERS)` — so LINTERS is now the single hand-authored
registration surface for generate-types discovery too. The move direction
(into linters.ts) is the only acyclic one: the discoverers already depend on
the pure parsers that live in linters.ts.
